Added Huds, Conections, phase 1

+++ src/main/java/net/cheto97/rpgcraftmod/modsystem/Customlevel.java
@@ -8,7 +8,7 @@ public class Customlevel {
 
     private int resetLevel = 0;
 
-    public double get(){
+    public int get(){
         return customlevel;
     }
     public double getMin(){
